requirements.txt
lxml
pandas
numpy
matplotlib
jupyter
scikit-learn
scikit-image
pytorch
torchvision